# .ai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wallet API",
    "description": "Sistema de Gestão Financeira",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sync/status": {
      "get": {
        "tags": [
          "Sincronização Nuvem"
        ],
        "summary": "Get Sync Pending Status",
        "description": "Retorna o status em tempo real de alterações pendentes para enviar e receber.",
        "operationId": "get_sync_pending_status_api_v1_sync_status_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "check_remote",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Consultar Google Sheets para verificar a Fila_Mobile",
              "default": true,
              "title": "Check Remote"
            },
            "description": "Consultar Google Sheets para verificar a Fila_Mobile"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/config": {
      "get": {
        "tags": [
          "Sincronização Nuvem"
        ],
        "summary": "Get Sync Configuration",
        "operationId": "get_sync_configuration_api_v1_sync_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConfigResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Sincronização Nuvem"
        ],
        "summary": "Update Sync Configuration",
        "operationId": "update_sync_configuration_api_v1_sync_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/sync/test": {
      "post": {
        "tags": [
          "Sincronização Nuvem"
        ],
        "summary": "Test Sync Connection",
        "operationId": "test_sync_connection_api_v1_sync_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SyncTriggerRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/sync/export": {
      "post": {
        "tags": [
          "Sincronização Nuvem"
        ],
        "summary": "Export Data To Sheets",
        "operationId": "export_data_to_sheets_api_v1_sync_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SyncTriggerRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/sync/import": {
      "post": {
        "tags": [
          "Sincronização Nuvem"
        ],
        "summary": "Import Data From Queue",
        "operationId": "import_data_from_queue_api_v1_sync_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SyncTriggerRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/sync/full": {
      "post": {
        "tags": [
          "Sincronização Nuvem"
        ],
        "summary": "Trigger Full Sync",
        "operationId": "trigger_full_sync_api_v1_sync_full_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SyncTriggerRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/sync/logs": {
      "get": {
        "tags": [
          "Sincronização Nuvem"
        ],
        "summary": "List Sync Logs",
        "operationId": "list_sync_logs_api_v1_sync_logs_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SyncLogResponse"
                  },
                  "title": "Response List Sync Logs Api V1 Sync Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sincronização Nuvem"
        ],
        "summary": "Clear Sync Logs",
        "operationId": "clear_sync_logs_api_v1_sync_logs_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncConfigResponse": {
        "properties": {
          "spreadsheet_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spreadsheet Id"
          },
          "spreadsheet_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spreadsheet Url"
          },
          "has_credentials": {
            "type": "boolean",
            "title": "Has Credentials",
            "default": false
          },
          "service_account_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Account Email"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "last_sync_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync Status"
          },
          "last_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Action"
          }
        },
        "type": "object",
        "title": "SyncConfigResponse"
      },
      "SyncConfigUpdate": {
        "properties": {
          "spreadsheet_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spreadsheet Id"
          },
          "credentials_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credentials Json"
          }
        },
        "type": "object",
        "title": "SyncConfigUpdate"
      },
      "SyncLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "items_imported": {
            "type": "integer",
            "title": "Items Imported"
          },
          "items_exported": {
            "type": "integer",
            "title": "Items Exported"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "action",
          "status",
          "items_imported",
          "items_exported",
          "message",
          "created_at"
        ],
        "title": "SyncLogResponse"
      },
      "SyncPendingDetails": {
        "properties": {
          "pending_transactions": {
            "type": "integer",
            "title": "Pending Transactions",
            "default": 0
          },
          "pending_categories": {
            "type": "integer",
            "title": "Pending Categories",
            "default": 0
          },
          "pending_items": {
            "type": "integer",
            "title": "Pending Items",
            "default": 0
          },
          "pending_accounts": {
            "type": "integer",
            "title": "Pending Accounts",
            "default": 0
          },
          "pending_contacts": {
            "type": "integer",
            "title": "Pending Contacts",
            "default": 0
          },
          "pending_debts": {
            "type": "integer",
            "title": "Pending Debts",
            "default": 0
          },
          "pending_budgets": {
            "type": "integer",
            "title": "Pending Budgets",
            "default": 0
          },
          "pending_attachments": {
            "type": "integer",
            "title": "Pending Attachments",
            "default": 0
          },
          "queue_rows": {
            "type": "integer",
            "title": "Queue Rows",
            "default": 0
          }
        },
        "type": "object",
        "title": "SyncPendingDetails"
      },
      "SyncResultResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "imported_from_queue": {
            "type": "integer",
            "title": "Imported From Queue",
            "default": 0
          },
          "exported_to_mirror": {
            "type": "integer",
            "title": "Exported To Mirror",
            "default": 0
          },
          "entity_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Entity Counts",
            "description": "Quantitativo de registros exportados por tabela"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "SyncResultResponse"
      },
      "SyncStatusResponse": {
        "properties": {
          "is_configured": {
            "type": "boolean",
            "title": "Is Configured",
            "default": false
          },
          "has_credentials": {
            "type": "boolean",
            "title": "Has Credentials",
            "default": false
          },
          "spreadsheet_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spreadsheet Id"
          },
          "spreadsheet_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spreadsheet Url"
          },
          "service_account_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Account Email"
          },
          "pending_send": {
            "type": "integer",
            "title": "Pending Send",
            "default": 0
          },
          "pending_receive": {
            "type": "integer",
            "title": "Pending Receive",
            "default": 0
          },
          "total_pending": {
            "type": "integer",
            "title": "Total Pending",
            "default": 0
          },
          "has_pending": {
            "type": "boolean",
            "title": "Has Pending",
            "default": false
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "last_sync_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync Status"
          },
          "last_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Action"
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncPendingDetails"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "SyncStatusResponse"
      },
      "SyncTestResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "spreadsheet_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spreadsheet Title"
          },
          "sheets_found": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sheets Found",
            "default": []
          },
          "service_account_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Account Email"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "SyncTestResponse"
      },
      "SyncTriggerRequest": {
        "properties": {
          "spreadsheet_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spreadsheet Id"
          }
        },
        "type": "object",
        "title": "SyncTriggerRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}